[Fix] 대중 평판 분석 내용을 실수로도 반환하도록 변경 1. DTO Map의 Integer을 Double로 수정 2. DTO 변경에 맞게 service 수정

+++ src/main/java/com/fifo/compasstep/reputationAnalysis/dto/response/ReputationDetailResponseDto.java
@@ -11,8 +11,8 @@ public record ReputationDetailResponseDto(
         Long historyId,
         String songTitle,
         String artistName,
-        Map<String, Integer> sentimentSummary, // {"positive":85,"negative":10,"neutral":5}
-        Map<String, Integer> emotionDetails,   // {"joy":50,"sadness":5,...}
+        Map<String, Double> sentimentSummary, // {"positive":85,"negative":10,"neutral":5}
+        Map<String, Double> emotionDetails,   // {"joy":50,"sadness":5,...}
         List<String> keywords,                 // ["달달하다","목소리",...]
         Instant createdAt
 ) {}
